+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.1.0 (2023-02-07)
 
+### 🌟 新功能
+范围|描述|commitId
+--|--|--
+ - | 完成页面 Header 与 Table 构建 | [bd001dd](https://github.com/picture-waterfall/picture-waterfall-admin/commit/bd001dd)
+
+
 ### 📝 文档
 范围|描述|commitId
 --|--|--